pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
dependencies = [
    #"rich",
    "pyyaml",
    "textual"
]